+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed `AI List Filter might not be working as expected` ([#220](https://github.com/architects-toolkit/SmartHopper/issues/220)).
+- Fixed limit to 4096 tokens for AI providers in settings dialog.
+- Fixed errors in validation methods of providers.
 - Fixed failing settings integrity check durinig initialization because providers were not loaded yet.
 
 ## [0.3.2-alpha] - 2025-06-15